backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

pydantic==2.8.2
pydantic-settings==2.4.0

httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0

python-dotenv==1.0.1
apscheduler==3.10.4
rapidfuzz==3.9.6
stripe==7.0.0
openai==1.54.0
PyPDF2==3.0.1
python-docx==1.1.0
python-magic-bin==0.4.14